Support only the escape sequences listed in The Semantics of K
Module Kore.Parser.CString is removed, because we do not support the full range
of C escape sequences (the module name was inaccurate). The respective escaping
and unescaping functions are moved to modules Kore.Parser.Lexeme and
Kore.Unparser. Module Kore.Parser.LexemeImpl is moved to Kore.Parser.Lexeme so
the compiler can detect unused code.
