+++ src/test/resources/expected/test-string-2.kore.golden
@@ -13,7 +13,7 @@ Definition
                             , applicationChildren =
                                 [ CofreeT
                                     { runCofreeT = Identity
-                                        { runIdentity = UnifiedMeta Null :< UnifiedMetaPattern (StringLiteralPattern (StringLiteral "\a"))
+                                        { runIdentity = UnifiedMeta Null :< UnifiedMetaPattern (StringLiteralPattern (StringLiteral "\x07"))
                                         }
                                     }
                                 ]
