+++ src/test/resources/expected/test-string-4.kore.golden
@@ -1,32 +1,5 @@
-Definition
-    { definitionAttributes =
-        Attributes 
-            [ KorePattern
-                { getKorePattern = CofreeT
-                    { runCofreeT = Identity
-                        { runIdentity = UnifiedObject Null :< UnifiedObjectPattern (ApplicationPattern Application
-                            { applicationSymbolOrAlias =
-                                SymbolOrAlias
-                                    { symbolOrAliasConstructor = (Id "key" AstLocationNone) :: Id Object
-                                    , symbolOrAliasParams = []
-                                    }
-                            , applicationChildren =
-                                [ CofreeT
-                                    { runCofreeT = Identity
-                                        { runIdentity = UnifiedMeta Null :< UnifiedMetaPattern (StringLiteralPattern (StringLiteral "\?77"))
-                                        }
-                                    }
-                                ]
-                            })
-                        }
-                    }
-                }
-            ]
-    , definitionModules =
-        [ Module
-            { moduleName = ModuleName "TEST-STRING-4"
-            , moduleSentences = []
-            , moduleAttributes = Attributes []
-            }
-        ]
-    }
+Parse error: ../../../test/resources//test-string-4.kore:1:11:
+  |
+1 | [key{}("\07777")]
+  |           ^
+expecting escape sequence
