+++ src/test/resources/test-string-1.kore
@@ -1,4 +1,4 @@
-[key{}("\?")]
+[key{}("?")]
 
 module TEST-STRING-1
