+++ src/test/resources/test-string-3.kore
@@ -1,4 +1,4 @@
-[key{}("\v")]
+[key{}("\u000B")]
 
 module TEST-STRING-3
